refactor user profile page
