+++ Dashboard/src/App.tsx
@@ -96,8 +96,11 @@ import ScheduledMaintenanceCustomFields from './Pages/Settings/ScheduledMaintena
 
 import ActiveIncidents from './Pages/Global/ActiveIncidents';
 import ProjectInvitations from './Pages/Global/ProjectInvitations';
-import UserProfile from './Pages/Global/UserProfile';
 
+// User Profile 
+import UserProfileOverview from './Pages/Global/UserProfile/Index';
+import UserProfilePicture from './Pages/Global/UserProfile/Picture';
+import UserProfilePassword from './Pages/Global/UserProfile/Password';
 // On Call Duty
 import OnCallDutyPage from './Pages/OnCallDuty/OnCallDuties';
 
@@ -293,7 +296,7 @@ const App: FunctionComponent = () => {
                         <NotOperationalMonitors
                             pageRoute={
                                 RouteMap[
-                                    PageMap.HOME_NOT_OPERATIONAL_MONITORS
+                                PageMap.HOME_NOT_OPERATIONAL_MONITORS
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -313,8 +316,8 @@ const App: FunctionComponent = () => {
                         <OngoingScheduledEvents
                             pageRoute={
                                 RouteMap[
-                                    PageMap
-                                        .HOME_ONGOING_SCHEDULED_MAINTENANCE_EVENTS
+                                PageMap
+                                    .HOME_ONGOING_SCHEDULED_MAINTENANCE_EVENTS
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -342,7 +345,7 @@ const App: FunctionComponent = () => {
                         <MonitorInoperational
                             pageRoute={
                                 RouteMap[
-                                    PageMap.MONITORS_INOPERATIONAL
+                                PageMap.MONITORS_INOPERATIONAL
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -384,7 +387,7 @@ const App: FunctionComponent = () => {
                         <MonitorViewStatusTimeline
                             pageRoute={
                                 RouteMap[
-                                    PageMap.MONITOR_VIEW_STATUS_TIMELINE
+                                PageMap.MONITOR_VIEW_STATUS_TIMELINE
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -401,7 +404,7 @@ const App: FunctionComponent = () => {
                         <MonitorIncidents
                             pageRoute={
                                 RouteMap[
-                                    PageMap.MONITOR_VIEW_INCIDENTS
+                                PageMap.MONITOR_VIEW_INCIDENTS
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -419,7 +422,7 @@ const App: FunctionComponent = () => {
                         <MonitorViewCustomFields
                             pageRoute={
                                 RouteMap[
-                                    PageMap.MONITOR_VIEW_CUSTOM_FIELDS
+                                PageMap.MONITOR_VIEW_CUSTOM_FIELDS
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -570,7 +573,7 @@ const App: FunctionComponent = () => {
                         <StatusPagesViewDelete
                             pageRoute={
                                 RouteMap[
-                                    PageMap.STATUS_PAGE_VIEW_DELETE
+                                PageMap.STATUS_PAGE_VIEW_DELETE
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -588,7 +591,7 @@ const App: FunctionComponent = () => {
                         <StatusPagesViewBranding
                             pageRoute={
                                 RouteMap[
-                                    PageMap.STATUS_PAGE_VIEW_BRANDING
+                                PageMap.STATUS_PAGE_VIEW_BRANDING
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -606,7 +609,7 @@ const App: FunctionComponent = () => {
                         <StatusPagesViewCustomHtmlCss
                             pageRoute={
                                 RouteMap[
-                                    PageMap.STATUS_PAGE_VIEW_CUSTOM_HTML_CSS
+                                PageMap.STATUS_PAGE_VIEW_CUSTOM_HTML_CSS
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -624,7 +627,7 @@ const App: FunctionComponent = () => {
                         <StatusPagesViewAdvancedOptions
                             pageRoute={
                                 RouteMap[
-                                    PageMap.STATUS_PAGE_VIEW_ADVANCED_OPTIONS
+                                PageMap.STATUS_PAGE_VIEW_ADVANCED_OPTIONS
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -642,7 +645,7 @@ const App: FunctionComponent = () => {
                         <StatusPageViewCustomFields
                             pageRoute={
                                 RouteMap[
-                                    PageMap.STATUS_PAGE_VIEW_CUSTOM_FIELDS
+                                PageMap.STATUS_PAGE_VIEW_CUSTOM_FIELDS
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -674,7 +677,7 @@ const App: FunctionComponent = () => {
                         <StatusPagesViewEmailSubscribers
                             pageRoute={
                                 RouteMap[
-                                    PageMap.STATUS_PAGE_VIEW_EMAIL_SUBSCRIBERS
+                                PageMap.STATUS_PAGE_VIEW_EMAIL_SUBSCRIBERS
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -682,7 +685,7 @@ const App: FunctionComponent = () => {
                     }
                 />
 
-<PageRoute
+                <PageRoute
                     path={
                         RouteMap[
                             PageMap.STATUS_PAGE_VIEW_AUTHENTICATION_SETTINGS
@@ -692,7 +695,7 @@ const App: FunctionComponent = () => {
                         <StatusPageViewAuthenticationSettings
                             pageRoute={
                                 RouteMap[
-                                    PageMap.STATUS_PAGE_VIEW_AUTHENTICATION_SETTINGS
+                                PageMap.STATUS_PAGE_VIEW_AUTHENTICATION_SETTINGS
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -710,7 +713,7 @@ const App: FunctionComponent = () => {
                         <StatusPageViewPrivateUser
                             pageRoute={
                                 RouteMap[
-                                    PageMap.STATUS_PAGE_VIEW_PRIVATE_USERS
+                                PageMap.STATUS_PAGE_VIEW_PRIVATE_USERS
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -728,7 +731,7 @@ const App: FunctionComponent = () => {
                         <StatusPagesViewSMSSubscribers
                             pageRoute={
                                 RouteMap[
-                                    PageMap.STATUS_PAGE_VIEW_SMS_SUBSCRIBERS
+                                PageMap.STATUS_PAGE_VIEW_SMS_SUBSCRIBERS
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -746,7 +749,7 @@ const App: FunctionComponent = () => {
                         <StatusPagesViewHeaderStyle
                             pageRoute={
                                 RouteMap[
-                                    PageMap.STATUS_PAGE_VIEW_HEADER_STYLE
+                                PageMap.STATUS_PAGE_VIEW_HEADER_STYLE
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -764,7 +767,7 @@ const App: FunctionComponent = () => {
                         <StatusPagesViewFooterStyle
                             pageRoute={
                                 RouteMap[
-                                    PageMap.STATUS_PAGE_VIEW_FOOTER_STYLE
+                                PageMap.STATUS_PAGE_VIEW_FOOTER_STYLE
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -782,7 +785,7 @@ const App: FunctionComponent = () => {
                         <StatusPagesViewNavBarStyle
                             pageRoute={
                                 RouteMap[
-                                    PageMap.STATUS_PAGE_VIEW_NAVBAR_STYLE
+                                PageMap.STATUS_PAGE_VIEW_NAVBAR_STYLE
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -800,7 +803,7 @@ const App: FunctionComponent = () => {
                         <StatusPagesViewWebhookSubscribers
                             pageRoute={
                                 RouteMap[
-                                    PageMap.STATUS_PAGE_VIEW_WEBHOOK_SUBSCRIBERS
+                                PageMap.STATUS_PAGE_VIEW_WEBHOOK_SUBSCRIBERS
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -818,7 +821,7 @@ const App: FunctionComponent = () => {
                         <StatusPagesViewEmbedded
                             pageRoute={
                                 RouteMap[
-                                    PageMap.STATUS_PAGE_VIEW_EMBEDDED
+                                PageMap.STATUS_PAGE_VIEW_EMBEDDED
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -836,7 +839,7 @@ const App: FunctionComponent = () => {
                         <StatusPagesViewResources
                             pageRoute={
                                 RouteMap[
-                                    PageMap.STATUS_PAGE_VIEW_RESOURCES
+                                PageMap.STATUS_PAGE_VIEW_RESOURCES
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -854,7 +857,7 @@ const App: FunctionComponent = () => {
                         <StatusPagesViewDomains
                             pageRoute={
                                 RouteMap[
-                                    PageMap.STATUS_PAGE_VIEW_DOMAINS
+                                PageMap.STATUS_PAGE_VIEW_DOMAINS
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -871,7 +874,7 @@ const App: FunctionComponent = () => {
                         <StatusPagesViewGroups
                             pageRoute={
                                 RouteMap[
-                                    PageMap.STATUS_PAGE_VIEW_GROUPS
+                                PageMap.STATUS_PAGE_VIEW_GROUPS
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -889,7 +892,7 @@ const App: FunctionComponent = () => {
                         <StatusPagesViewAnnouncement
                             pageRoute={
                                 RouteMap[
-                                    PageMap.STATUS_PAGE_VIEW_ANNOUNCEMENTS
+                                PageMap.STATUS_PAGE_VIEW_ANNOUNCEMENTS
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -957,7 +960,7 @@ const App: FunctionComponent = () => {
                         <IncidentViewStateTimeline
                             pageRoute={
                                 RouteMap[
-                                    PageMap.INCIDENT_VIEW_STATE_TIMELINE
+                                PageMap.INCIDENT_VIEW_STATE_TIMELINE
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -974,7 +977,7 @@ const App: FunctionComponent = () => {
                         <IncidentInternalNote
                             pageRoute={
                                 RouteMap[
-                                    PageMap.INCIDENT_INTERNAL_NOTE
+                                PageMap.INCIDENT_INTERNAL_NOTE
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -992,7 +995,7 @@ const App: FunctionComponent = () => {
                         <IncidentViewCustomFields
                             pageRoute={
                                 RouteMap[
-                                    PageMap.INCIDENT_VIEW_CUSTOM_FIELDS
+                                PageMap.INCIDENT_VIEW_CUSTOM_FIELDS
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -1026,7 +1029,7 @@ const App: FunctionComponent = () => {
                         <ScheduledMaintenanceEvents
                             pageRoute={
                                 RouteMap[
-                                    PageMap.SCHEDULED_MAINTENANCE_EVENTS
+                                PageMap.SCHEDULED_MAINTENANCE_EVENTS
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -1044,7 +1047,7 @@ const App: FunctionComponent = () => {
                         <OngoingScheduledMaintenanceEvents
                             pageRoute={
                                 RouteMap[
-                                    PageMap.ONGOING_SCHEDULED_MAINTENANCE_EVENTS
+                                PageMap.ONGOING_SCHEDULED_MAINTENANCE_EVENTS
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -1062,7 +1065,7 @@ const App: FunctionComponent = () => {
                         <ScheduledMaintenanceEventView
                             pageRoute={
                                 RouteMap[
-                                    PageMap.SCHEDULED_MAINTENANCE_VIEW
+                                PageMap.SCHEDULED_MAINTENANCE_VIEW
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -1080,8 +1083,8 @@ const App: FunctionComponent = () => {
                         <ScheduledMaintenanceEventsViewCustomFields
                             pageRoute={
                                 RouteMap[
-                                    PageMap
-                                        .SCHEDULED_MAINTENANCE_VIEW_CUSTOM_FIELDS
+                                PageMap
+                                    .SCHEDULED_MAINTENANCE_VIEW_CUSTOM_FIELDS
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -1099,7 +1102,7 @@ const App: FunctionComponent = () => {
                         <ScheduledMaintenanceEventViewDelete
                             pageRoute={
                                 RouteMap[
-                                    PageMap.SCHEDULED_MAINTENANCE_VIEW_DELETE
+                                PageMap.SCHEDULED_MAINTENANCE_VIEW_DELETE
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -1117,8 +1120,8 @@ const App: FunctionComponent = () => {
                         <ScheduledMaintenanceEventViewStateTimeline
                             pageRoute={
                                 RouteMap[
-                                    PageMap
-                                        .SCHEDULED_MAINTENANCE_VIEW_STATE_TIMELINE
+                                PageMap
+                                    .SCHEDULED_MAINTENANCE_VIEW_STATE_TIMELINE
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -1136,7 +1139,7 @@ const App: FunctionComponent = () => {
                         <ScheduledMaintenanceEventInternalNote
                             pageRoute={
                                 RouteMap[
-                                    PageMap.SCHEDULED_MAINTENANCE_INTERNAL_NOTE
+                                PageMap.SCHEDULED_MAINTENANCE_INTERNAL_NOTE
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -1154,7 +1157,7 @@ const App: FunctionComponent = () => {
                         <ScheduledMaintenanceEventPublicNote
                             pageRoute={
                                 RouteMap[
-                                    PageMap.SCHEDULED_MAINTENANCE_PUBLIC_NOTE
+                                PageMap.SCHEDULED_MAINTENANCE_PUBLIC_NOTE
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -1216,7 +1219,7 @@ const App: FunctionComponent = () => {
                         <SettingsMonitors
                             pageRoute={
                                 RouteMap[
-                                    PageMap.SETTINGS_MONITORS_STATUS
+                                PageMap.SETTINGS_MONITORS_STATUS
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -1234,7 +1237,7 @@ const App: FunctionComponent = () => {
                         <SettingsIncidents
                             pageRoute={
                                 RouteMap[
-                                    PageMap.SETTINGS_INCIDENTS_STATE
+                                PageMap.SETTINGS_INCIDENTS_STATE
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -1252,7 +1255,7 @@ const App: FunctionComponent = () => {
                         <SettingsScheduledMaintenanceState
                             pageRoute={
                                 RouteMap[
-                                    PageMap.SETTINGS_SCHEDULED_MAINTENANCE_STATE
+                                PageMap.SETTINGS_SCHEDULED_MAINTENANCE_STATE
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -1280,7 +1283,7 @@ const App: FunctionComponent = () => {
                         <SettingsIncidentSeverity
                             pageRoute={
                                 RouteMap[
-                                    PageMap.SETTINGS_INCIDENTS_SEVERITY
+                                PageMap.SETTINGS_INCIDENTS_SEVERITY
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -1350,7 +1353,7 @@ const App: FunctionComponent = () => {
                         <MonitorCustomFields
                             pageRoute={
                                 RouteMap[
-                                    PageMap.SETTINGS_MONITOR_CUSTOM_FIELDS
+                                PageMap.SETTINGS_MONITOR_CUSTOM_FIELDS
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -1368,7 +1371,7 @@ const App: FunctionComponent = () => {
                         <StatusPageCustomFields
                             pageRoute={
                                 RouteMap[
-                                    PageMap.SETTINGS_STATUS_PAGE_CUSTOM_FIELDS
+                                PageMap.SETTINGS_STATUS_PAGE_CUSTOM_FIELDS
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -1386,8 +1389,8 @@ const App: FunctionComponent = () => {
                         <ScheduledMaintenanceCustomFields
                             pageRoute={
                                 RouteMap[
-                                    PageMap
-                                        .SETTINGS_SCHEDULED_MAINTENANCE_CUSTOM_FIELDS
+                                PageMap
+                                    .SETTINGS_SCHEDULED_MAINTENANCE_CUSTOM_FIELDS
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -1405,7 +1408,7 @@ const App: FunctionComponent = () => {
                         <IncidentCustomFields
                             pageRoute={
                                 RouteMap[
-                                    PageMap.SETTINGS_INCIDENT_CUSTOM_FIELDS
+                                PageMap.SETTINGS_INCIDENT_CUSTOM_FIELDS
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -1435,7 +1438,7 @@ const App: FunctionComponent = () => {
                         <SettingsInvoices
                             pageRoute={
                                 RouteMap[
-                                    PageMap.SETTINGS_BILLING_INVOICES
+                                PageMap.SETTINGS_BILLING_INVOICES
                                 ] as Route
                             }
                             currentProject={selectedProject}
@@ -1508,10 +1511,30 @@ const App: FunctionComponent = () => {
 
                 {/* Global Routes */}
                 <PageRoute
-                    path={RouteMap[PageMap.USER_PROFILE]?.toString() || ''}
+                    path={RouteMap[PageMap.USER_PROFILE_PICTURE]?.toString() || ''}
                     element={
-                        <UserProfile
-                            pageRoute={RouteMap[PageMap.USER_PROFILE] as Route}
+                        <UserProfilePicture
+                            pageRoute={RouteMap[PageMap.USER_PROFILE_PICTURE] as Route}
+                            currentProject={selectedProject}
+                        />
+                    }
+                />
+
+                <PageRoute
+                    path={RouteMap[PageMap.USER_PROFILE_OVERVIEW]?.toString() || ''}
+                    element={
+                        <UserProfileOverview
+                            pageRoute={RouteMap[PageMap.USER_PROFILE_OVERVIEW] as Route}
+                            currentProject={selectedProject}
+                        />
+                    }
+                />
+
+                <PageRoute
+                    path={RouteMap[PageMap.USER_PROFILE_PASSWORD]?.toString() || ''}
+                    element={
+                        <UserProfilePassword
+                            pageRoute={RouteMap[PageMap.USER_PROFILE_PASSWORD] as Route}
                             currentProject={selectedProject}
                         />
                     }
