+++ Dashboard/src/Utils/PageMap.ts
@@ -113,7 +113,9 @@ enum PageMap {
 
     // Misc
     LOGOUT = 'LOGOUT',
-    USER_PROFILE = 'USER_PROFILE',
+    USER_PROFILE_OVERVIEW = 'USER_PROFILE_OVERVIEW',
+    USER_PROFILE_PASSWORD = 'USER_PROFILE_PASSWORD',
+    USER_PROFILE_PICTURE = 'USER_PROFILE_PICTURE',
     ACTIVE_INCIDENTS = 'ACTIVE_INCIDENTS',
     PROJECT_INVITATIONS = 'PROJECT_INVITATIONS',
